requirements.txt
flask==2.0.1
werkzeug==2.0.1
redis==3.5.3
mysql-connector-python==8.0.26
prometheus-client==0.11.0
gunicorn==20.1.0 